+++ src/components/BrowseResultsList/constants.js
@@ -10,9 +10,9 @@ export const VISIBLE_COLUMNS_MAP = {
   [browseModeOptions.LOCAL]: ['callNumber', 'title', 'numberOfTitles'],
   [browseModeOptions.NATIONAL_LIBRARY_OF_MEDICINE]: ['callNumber', 'title', 'numberOfTitles'],
   [browseModeOptions.OTHER]: ['callNumber', 'title', 'numberOfTitles'],
-  [browseModeOptions.CLASSIFICATION_ALL]: ['classificationNumber', 'contributors', 'numberOfTitles'],
-  [browseModeOptions.DEWEY_CLASSIFICATION]: ['classificationNumber', 'contributors', 'numberOfTitles'],
-  [browseModeOptions.LC_CLASSIFICATION]: ['classificationNumber', 'contributors', 'numberOfTitles'],
+  [browseModeOptions.CLASSIFICATION_ALL]: ['classificationNumber', 'title', 'contributors', 'numberOfTitles'],
+  [browseModeOptions.DEWEY_CLASSIFICATION]: ['classificationNumber', 'title', 'contributors', 'numberOfTitles'],
+  [browseModeOptions.LC_CLASSIFICATION]: ['classificationNumber', 'title', 'contributors', 'numberOfTitles'],
   [browseModeOptions.SUPERINTENDENT]: ['callNumber', 'title', 'numberOfTitles'],
   [browseModeOptions.CONTRIBUTORS]: ['contributor', 'contributorType', 'relatorTerm', 'numberOfTitles'],
 };
@@ -71,15 +71,21 @@ export const COLUMNS_WIDTHS = {
   },
   [browseModeOptions.CLASSIFICATION_ALL]: {
     classificationNumber: '20%',
+    title: '35%',
     contributors: '30%',
+    numberOfTitles: '15%',
   },
   [browseModeOptions.DEWEY_CLASSIFICATION]: {
     classificationNumber: '20%',
+    title: '35%',
     contributors: '30%',
+    numberOfTitles: '15%',
   },
   [browseModeOptions.LC_CLASSIFICATION]: {
     classificationNumber: '20%',
+    title: '35%',
     contributors: '30%',
+    numberOfTitles: '15%',
   },
   [browseModeOptions.CONTRIBUTORS]: {
     contributor: '50%',