+++ CHANGELOG.md
@@ -27,6 +27,7 @@ UIIN-3437.
 * ECS: Provide correct tenant to get member tenant holdings for instance. Fixes UIIN-3447.
 * Fix hotkey handlers for Instance details. Fixes UIIN-3448.
 * Classification browse | Add Contributors to results list. Refs UIIN-3368.
+* Classification browse | Add Title to results list. Refs UIIN-3367.
 
 ## [13.0.8](https://github.com/folio-org/ui-inventory/tree/v13.0.8) (2025-07-16)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v13.0.7...v13.0.8)